:recycle: refactor: substituicao para o padrao ES Modules #167
