+++ backend/package.json
@@ -2,7 +2,7 @@
     "name": "api",
     "version": "1.0.0",
     "main": "server.js",
-    "type": "commonjs",
+    "type": "module",
     "scripts": {
         "test": "jest --forceExit",
         "start": "nodemon server.js"
