+++ backend/server.js
@@ -1,13 +1,15 @@
-const app = require('./app');
-const connectDb = require('./db');
-require('dotenv').config();
+import 'dotenv/config';
+import app from './src/app.js';
+import connectDb from "./src/db.js";
+
+const PORT = 3000;
 
 // Conexão com o banco e com o servidor
 connectDb()
     .then(data => {
         console.log(' >> Banco de dados conectado com sucesso:\n')
-        app.listen(3000, () => {
-            console.log(`>> Servidor rodando na porta http://localhost:${3000}/home\n`)
+        app.listen(PORT, () => {
+            console.log(`>> Servidor rodando na porta http://localhost:${PORT}/home\n`)
         }).on('error', err =>
             console.log('Erro ao ligar o servidor:\n', err))
     })
